topology2: Add classes for DAPM routes
Add a connection class that will be used for creating DAPM routes
between widgets. A connection object can be of type "graph"
that connects two widgets in a pipeline or it can be of
type "endpoint" that connects a pipeline endpoint to a DAI
endpoint or two pipeline endpoints.

Also, add some helper classes for endpoints and source/sink widgets
that will be used to define the connection object.

Signed-off-by: Ranjani Sridharan <ranjani.sridharan@linux.intel.com>