topology2: Add pipeline classes
Add 2 pipeline classes for Volume playback and Volume Capture.
Each of these classes includes the widget objects, their connections
and define attributes for the pipeline.

A volume playback pipeline can be instanted as follows:
Object.pipeline-volume-playback."2" {
	channels	2
	channels_min	2
	channels_max	2
	rate		48000
	rate_min	48000
	rate_max	48000
	format			"s32le"
}

Every pipeline class is associated with a pipeline widget, so
add the class definition for the pipeline widget as well.

Signed-off-by: Ranjani Sridharan <ranjani.sridharan@linux.intel.com>