Topology2: rename payload_with_output_fmt token and extend its use
The output format is not the only optional content in
the initialization payload, but the current used boolean
token payload_with_output_fmt is specific to output
format, and is not extensible.

This patch renames the token name to init_payload_format,
and the optional output format only takes bit 0, bits 1~31 are
reserved for future use.

Signed-off-by: Chao Song <chao.song@linux.intel.com>
Signed-off-by: Ranjani Sridharan <ranjani.sridharan@linux.intel.com>